requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.40.6
requests==2.32.3
httpx==0.27.2